requirements.txt
openai-whisper
easyocr
torch
torchvision
torchaudio
opencv-python
numpy
